Correções na recombinação do AG.

+++ trunk/src/AG/operadores_geneticos.c
@@ -3,7 +3,7 @@
 #include "parametros_ag.h"
 #include <stdio.h>
 
-void mutacao(individuo * p, double chance){
+void mutacao(individuo * p, float chance){
 
     int i;
 
@@ -11,6 +11,7 @@ void mutacao(individuo * p, double chance){
 
         //gera um número entre 0 e 1
         float aleatorio = (float)rand()/RAND_MAX;
+
 
         if (aleatorio<chance) {
            p->genotipo[i] = (p->genotipo[i] +1)%2;
@@ -36,10 +37,10 @@ void crossover_um_ponto(individuo *pai1, individuo *pai2, individuo *filho1, ind
     }
 }
 
-void recombinacao(individuo *pai1, individuo*pai2, individuo*filho1, individuo*filho2, double chance){
+void recombinacao(individuo *pai1, individuo*pai2, individuo*filho1, individuo*filho2, float chance){
 
     //gera um número entre 0 e 1
-    double aleatorio = (float)rand()/RAND_MAX;
+    float aleatorio = (float)rand()/RAND_MAX;
 
     if (aleatorio<chance) {
         crossover_um_ponto(pai1, pai2,filho1, filho2);
