+++ trunk/src/AG/operadores_geneticos.h
@@ -4,7 +4,7 @@
 #include "parametros_ag.h"
 #include "representacao.h"
 
-void mutacao(individuo* individuo, double chance);
+void mutacao(individuo* individuo, float chance);
 void crossover_um_ponto(individuo *pai1, individuo*pai2,individuo *filho1, individuo *filho2);
 void recombinacao(individuo *pai1, individuo *pai2, individuo *filho1, individuo *filho2, float chance);
